% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quality.R
\name{check_quality}
\alias{check_quality}
\alias{check_quality.both_zero_na}
\alias{check_quality.na_refs}
\alias{check_quality.both_not_zero}
\alias{check_quality.not_na}
\alias{check_quality.values_unique_comb}
\title{Quality checks}
\usage{
check_quality(type, ...)

\method{check_quality}{both_zero_na}(
  raw_dat = NULL,
  checked = NULL,
  type = type,
  filename = NULL,
  dir_log = NULL,
  ...
)

\method{check_quality}{na_refs}(
  raw_dat = NULL,
  checked = NULL,
  refs = NULL,
  type = type,
  filename = NULL,
  dir_log = NULL,
  ...
)

\method{check_quality}{both_not_zero}(
  raw_dat = NULL,
  checked = NULL,
  refs = NULL,
  type = type,
  filename = NULL,
  dir_log = NULL,
  ...
)

\method{check_quality}{not_na}(
  raw_dat = NULL,
  checked = NULL,
  type = type,
  filename = NULL,
  dir_log = NULL,
  ...
)

\method{check_quality}{values_unique_comb}(
  raw_dat = NULL,
  checked = NULL,
  refs = NULL,
  values = NULL,
  type = type,
  filename = NULL,
  dir_log = NULL,
  ...
)
}
\arguments{
\item{type}{Type of quality check to be performed, either both_zero_na, na_refs, both_not_zero, not_na, smaller, bigger, equal, inside_bbox, inside_timespan}

\item{raw_dat}{Dataframe. Dataset on which the quality assessment is performed}

\item{checked}{A vector of the name(s) of column(s) whose quality is checked.}

\item{filename}{Filename of the raw dataset related to the metadata}

\item{refs}{A vector of the name(s) of the reference column(s) to which the checked_cols will be compared}
}
\value{
NULL or vector of integer representing the lines that contains error
}
\description{
Set of functions to check quality of raw data before they get converted to a .csv file.
}
\section{both_zero_na}{

}

\section{na_refs}{

}

\section{both_not_zero}{

}

\section{not_na}{

}

\section{values_unique_comb}{

}

\section{smaller}{

}

\section{larger}{

}

\section{equal}{

}

\section{inside_bbox}{

}

\section{inside_timespan}{

}

