% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curating.R
\name{curating}
\alias{curating}
\title{Curating raw data}
\usage{
curating(metadat_path)
}
\arguments{
\item{metadat_path}{The path to metadata file}
}
\value{
Saves curated data in specified folder
}
\description{
This function uses metadata .yaml file to curate raw data. It makes quality assessments based on the "quality" section of the metadata file. It also convert date in "yyyy-mm-dd" format, time in "HH:MM:SS" format and lat/lon in decimal degree format. It also removes whitespace and uppercase in column names.
}
