% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertion.R
\name{format_raw_dat}
\alias{format_raw_dat}
\title{Converting and formatting date, time, geom, units columns}
\usage{
format_raw_dat(raw_dat, metadat_convertion)
}
\arguments{
\item{raw_dat}{Dataframe. Dataset to format}

\item{metadat_convertion}{List with information about date, time and lat/lon columns names and format, and the measurements columns that contain unit in their name}
}
\value{
formatted dataset
}
\description{
This set of functions convert and format data in five ways. It format date to "YYYY-mm-dd", it format time to "HH:MM:SS", it converts lat/lon to decimal degree, is create units columns if the unit of a measurements is initially given in the names of the measurement columns, and it remove every whitespace and capital letters from the column name. It's basically a minimal convertion and formatting to clean the datasets.
}
