% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_assessment.R
\name{quality_assessment}
\alias{quality_assessment}
\title{Quality assessments}
\usage{
quality_assessment(dat, assertions)
}
\arguments{
\item{assertions}{List. List of data quality assertions to test using quality_test.* functions. This list must include the test name, the column(s) on which to perform the test and extra parameters to use in the quality_test.* function.}

\item{x}{Dataframe. Dataset on which the quality assessment is performed}
}
\value{
List of two dataframe. Output: rows that passed the assessment and Error: rows that failed at least one test.
}
\description{
Quality assessments
}
