% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dat.R
\name{read_dat}
\alias{read_dat}
\title{Import raw data from raw_data folder}
\usage{
read_dat(x = NULL, format = NULL, sheet = NULL)
}
\arguments{
\item{x}{(character) Name of the file to import}

\item{format}{(character) Format of the file to import}

\item{sheet}{(character) If format if 'xls' or 'xlsx', specify which sheet(s) to import}
}
\value{
data.frame or a list for metadata
}
\description{
Function to import raw data from different formats from the raw_data
}
