% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quality.R
\name{quality_assessment}
\alias{quality_assessment}
\title{Quality checks}
\usage{
quality_assessment(raw_dat, metadat_quality = NULL, filename = NULL)
}
\arguments{
\item{raw_dat}{Dataframe. Dataset on which the quality assessments is performed}

\item{metadat_quality}{A list of metadata quality assessments to perform on the dataset}

\item{filename}{Filename of the raw dataset related to the metadata}
}
\value{
Dataframe with observations that passed the quality assessments
}
\description{
Set of functions to check quality of raw data before they get converted to a .csv file.
}
